@@ -4,6 +4,6 @@ export default create({
   base: 'light',
   brandTitle: 'NiceNode UI Components',
   brandUrl: 'https://nndesign.netlify.app/',
-  brandImage: 'https://www.nicenode.xyz/img/icon.png',
+  brandImage: 'https://www.nicenode.xyz/favicon.png',
   brandTarget: '_self',
 });
