@@ -8,7 +8,7 @@ export const parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
   controls: {
     matchers: {
-      color: /(background|color)$/i,
+      // color: /(background|color)$/i,
       date: /Date$/,
     },
   },
