@@ -23,7 +23,8 @@ export type NiceNodeRpcTranslation =
   | 'eth-l1-beacon'
   | 'eth-l2-starknet'
   | 'eth-l2-arbitrum'
-  | 'eth-l2-consensus';
+  | 'eth-l2-consensus'
+  | 'farcaster-l1';
 // | {
 //     peerCount?: RpcTranslation;
 //     sync?: RpcTranslation;
