@@ -74,7 +74,6 @@ const NodeScreen = () => {
       pollingInterval,
     },
   );
-
   const qIsPodmanRunning = useGetIsPodmanRunningQuery(null, {
     pollingInterval: 15000,
   });
@@ -201,6 +200,8 @@ const NodeScreen = () => {
       rpcTranslation === 'eth-l1-beacon'
     ) {
       latestBlockNum = parseFloat(slotNumber);
+    } else if (rpcTranslation === 'farcaster-l1') {
+      latestBlockNum = qLatestBlock.data;
     }
 
     const syncedBlock =
@@ -301,61 +302,23 @@ const NodeScreen = () => {
     return info;
   };
 
-  const formatVersion = (version: string | undefined, name: string) => {
+  const formatVersion = (version: string | undefined) => {
     if (!version) {
       return t('IdentifyingVersion');
     }
-    const capitalize = (s: string) =>
-      (s && s[0].toUpperCase() + s.slice(1)) || '';
 
-    let regex;
-    switch (name) {
-      case 'geth':
-      case 'op-geth':
-      case 'op-node':
-        regex = /Geth\/v(\d+\.\d+\.\d+)/;
-        break;
-      case 'reth':
-        regex = /Reth\/v(\d+\.\d+\.\d+)/;
-        break;
-      case 'besu':
-        regex = /besu\/v(\d+\.\d+\.\d+)/;
-        break;
-      case 'erigon':
-        regex = /(\d+\.\d+\.\d+)-dev/;
-        break;
-      case 'nethermind':
-      case 'lighthouse':
-        // eslint-disable-next-line no-useless-escape
-        regex = new RegExp(`${capitalize(name)}\/v(\\d+\\.\\d+\\.\\d+)`);
-        break;
-      case 'lodestar':
-        regex = /Lodestar\/v(\d+\.\d+\.\d+)/;
-        break;
-      case 'prysm':
-        regex = /prysm\/v(\d+\.\d+\.\d+)/;
-        break;
-      case 'teku':
-        regex = /teku\/v(\d+\.\d+\.\d+)/;
-        break;
-      case 'nimbus':
-        regex = /Nimbus\/v(\d+\.\d+\.\d+)/;
-        break;
-      default:
-        console.error(
-          `Can not parse node version string. Node name not found: ${name}`,
-        );
-        return version; // At least, return the unformatted version string
-    }
+    // At least, return the unformatted version string
+    let versionString = version;
 
-    const match = version.match(regex);
+    // Don't match -stable or -<commit-hash> that many nodes use
+    const genericVersionRegex = /v\d+\.\d+\.\d+(-[alpha|beta|edge]+)?/;
+    const match = version.match(genericVersionRegex);
 
     if (match) {
-      const versionString = `v${match[1]}`;
-      return versionString;
+      versionString = match[0];
     }
-    console.error(`No version number found for ${name}`);
-    return '';
+
+    return versionString;
   };
 
   const clientName = spec.specId.replace('-beacon', '');
@@ -368,7 +331,7 @@ const NodeScreen = () => {
     name: clientName as NodeBackgroundId,
     screenType: 'client',
     rpcTranslation: spec.rpcTranslation,
-    version: formatVersion(nodeVersionData, clientName),
+    version: formatVersion(nodeVersionData),
     info: formatSpec(
       spec.category,
       // selectedNode?.config?.configValuesMap?.network ?? '',
